requirements.txt
asgiref==3.7.2
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
prompt-toolkit==3.0.39
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
pyvmomi==8.0.2.0
six==1.16.0
sqlparse==0.4.4
wcwidth==0.2.9
paramiko~=3.3.1
